Fixed Surface bug, missing mipmap creation
Fixed Surface bug, missing mipmap creation

This was a truly annoying bug, as no error was reported at any point, be that by
the CPU or the GPU. Texture data just ended up being vec4(0) when accessed in
the fragment shader. In spite of this, the glDEbugger showed that the texture is
if everything was normal. Apparently, if you tell OpenGL that the
GL_TEXTURE_MIN_FILTER is GL_LINEAR_MIPMAP_LINEAR, you better make sure to actually
generate the mipmaps with glGenerateMipmap(GL_TEXTURE_2D)

Anisotropic filtering is also disabled on OpenGL ES

    src/graphics/Surface.cpp