Communication between flite and SDL
Communication between flite and SDL

Filte creates mono WAV data at samplerate 16000, 16 bit.

This is different from the current SDL_mixer configuration (22050 samplerate,
16bit, stereo), and as such needs to transform the data prior to playback.

The best solution would be to modify flite library to produce
output samples matching the SDL configuration.

The next best solution is to manually transform the samples based on existing
knowledge of the WAV data.

The third best solution (and the one currently implemented) is to
create a WAV RIFF header on top of the WAV data, and hand that over to
SDL_mixer, which reads it and transforms the data to conform to the output
format.

    src/audio/FliteUtil.cpp (+)
    src/audio/FliteUtil.h   (+)
    src/audio/SoundChunk.cpp
    src/audio/SoundChunk.h