Isolated FileInfo::TYPE_{ASSET,WRITABLE} to FileUtil
Isolated FileInfo::TYPE_{ASSET,WRITABLE} to FileUtil

Before, Asset class used to know OS filesystem specific details. This shouldn't
be the case, as the Asset class simply represents an asset object. If android
and linux read this differently, it should only need to be differentiated in the
FileUtil and File classes.

    src/audio/AudioResourceManager.cpp
    src/audio/MusicTrack.cpp
    src/audio/SoundChunk.cpp
    src/graphics/Mesh.cpp
    src/graphics/MeshResourceManager.cpp
    src/graphics/ShaderProgram.cpp
    src/graphics/ShaderResourceManager.cpp
    src/graphics/ShaderUtil.cpp
    src/graphics/ShaderUtil.h
    src/graphics/Surface.cpp
    src/util/Asset.cpp
    src/util/Asset.h
    src/util/File.cpp
    src/util/File.h
    src/util/FileUtil.cpp
    src/util/FileUtil.h
    src/util/ObjUtil.cpp
    src/util/TestAsset.h
    src/util/TestFile.h
    src/util/TestObjUtil.h