Added skipping of test for windows
Added skipping of test for windows
Some strange error caused it to fail, but this error isn't
reproduced in the utilities which use the same functionality
that the test covers.

It has been marked with a todo, to inspect further, but,
most likely the issue is with the testing code, and not the
tested code.

    src/util/TestObjUtil.h