Replaced Umundo with lower level ZMQ/CZMQ
Replaced Umundo with lower level ZMQ/CZMQ

The reason for this was multiple bugs with the umundo implementation, and
over-engineering of the library (if every second class has a shared_from_this,
something is not right). The decision was instead to build on the core library
of ZMQ which is much more robust.

Unfortunately complications in getting mDNSResponder to work for the discovery
lead to using zbeacon from the czmq library. This is a far less elegant way of
doing network discovery, since broadcast packages are spammed continuously.

However, whenever the time allows, replacing czmq with mDNSResponder should
not affect the rest of the system.

The discovery is still rudimentary. This commit is mostly build system update for
linux/windows/android of zmq and czmq, and removing deprecated umundo stuff.

Removed umundo files:
    external/umundo/*                            (-)
    utils/build_scripts/umundo/*                 (-)
    utils/scripts/prepare_umundo_dependencies.sh (-)

Updated files
    README.md
    SConstruct
    src/Android.mk
    src/SConscript
    src/SConscript_tests
    src/net/NetDiscovery.cpp
    src/net/NetDiscovery.h
    utils/linux_build.py
    utils/scripts/crosscompile_libs.sh
    utils/scripts/initialize_project.sh
    utils/win_build.py

Added netlibs files
    external/netlibs/Android.mk                          (+)
    external/netlibs/SConscript                          (+)
    src/net/Zmq.h                                        (+)
    src/util/TestUuid.h                                  (+)
    src/util/Uuid.cpp                                    (+)
    src/util/Uuid.h                                      (+)
    utils/build_scripts/netlibs/Android_czmq.mk          (+)
    utils/build_scripts/netlibs/Android_zmq.mk           (+)
    utils/build_scripts/netlibs/SConscript_czmq          (+)
    utils/build_scripts/netlibs/SConscript_zeromq        (+)
    utils/build_scripts/netlibs/android-build-czmq.sh    (+)
    utils/build_scripts/netlibs/android-build-zeromq.sh  (+)
    utils/build_scripts/netlibs/czmq_platform_android.h  (+)
    utils/build_scripts/netlibs/czmq_platform_linux.h    (+)
    utils/build_scripts/netlibs/czmq_platform_windows.h  (+)
    utils/build_scripts/netlibs/zmq_platform_android.hpp (+)
    utils/build_scripts/netlibs/zmq_platform_linux.hpp   (+)
    utils/build_scripts/netlibs/zmq_platform_windows.hpp (+)
    utils/scripts/prepare_netlibs.sh                     (+)
    utils/patches/czmq-3.0.0.mingw.patch                 (+)